requirements.txt
django
djangorestframework
django-filter
django-cors-headers
requests
drf-yasg
psycopg2-binary
pyyaml
